sonar issues fix

+++ src/main/java/com/rte_france/antares/datamanager_back/util/timeseries_manager/TimeSeriesReader.java
@@ -140,7 +140,7 @@ public final class TimeSeriesReader {
 
       // Load data into a matrix
       int actualRowCount = Math.clamp(
-              sheet.getLastRowNum() - firstRow.getRowNum(),
+              (long) sheet.getLastRowNum() - firstRow.getRowNum(),
               0,
               ROW_COUNT
       );
